Add NFS volume to secbox
suse-cmd was an experiment and it is now named secbox.
Podman is now used to manage the NFS volume, hence the container engine
now requires root priviledges. Processes within the container will keep
running as unpriviledged user.
It is very important to run this script with SUDO (sudo -E), this way the
user within the container is your normal user in the host, otherwise a
root user will be configured within the container and it will be mapped to
the real root user in the host. Check the aliases in my .zshrc to see how
I use this script.