Automatically mount and umount NFS shares
Having NFS shares mounted since the first secbox execution could lead
to unwanted scenarios. If a user logout from the VPN while the shares
are still mounted, connections to NFS servers hang and a reboot is
required, this is very well described here [0].
Whit this commit NFS shares are mounted only when the --nfs flag is set
and automatically umounted after process termination. This should
mitigate the risk above mentioned, but users have to know whe use the
--nfs flag. A good example could be: alias tel="secbox --nfs tel"
since tel is an internal tool located in /mounts/internal-tools.

[0] https://www.suse.com/support/kb/doc/?id=000019722