Automatically manage image update
The container image is now automatically built by IBS whenever a new
release is issued in the GitLab repo [0]. Secbox detects newly available
images in the internal registry [1] and asks the user if he wants to
update the local container.

TODO: allow the user to suppress the update notification in case he
doesn't want to update the container, otherwise this will appear at any
issued command.

[0] https://gitlab.suse.de/security/secbox-image
[1] https://registry.suse.de