tldr-pages installation and design change
tldr [0] is a collection of community-maintained help pages for
command-line tools. I chose the sh client [1] for portability and
light deps. While installing it as git submodule I realized adding its
repo to my PATH (as I used to do for another tool: secbox) might result
in unwanted behaviors when additional binaries exist in the same
cloned repo, and this is the case of tldr-sh-client and the `test`
binary [2]. Luckly test is also a zsh builtin, so it would not have
been picked up in case of execution... but in future with other cases I
may not be that lucky. Hence I moved to a better design leveraging
linked files. For that, I will use ~/.local/bin/repos as bucket for
tools I will install as git submodules which don't require to be in a
more convinient place (eg. zsh plugins).

[0] https://github.com/tldr-pages/tldr
[1] https://github.com/raylee/tldr-sh-client
[2] https://github.com/raylee/tldr-sh-client/blob/master/test