Slack and Teams as fake native clients
This commit allows to execute Slack and Teams respectively with
commands: slack, teams. Those are executed in Firefox or Chromium
instances which leverages dedicated profiles and are customized to
better fit the kiosk mode, making those webapps frameless.

Plus, weechat is now a wrapper to run the weechat client from its
own alacritty instance. The --class option is used to assign an
a custom app_id to the window.