Flatpak: Thunderbird doesn't need access to .gnupg
Since TB 78 relies on built-in encryption standard as replacement of
Enigmail [0]. Now TB manages GPG keys itself and stores them in the user
profile folder [1], which means that access to ~/.gnupg is no longer
required.

Access to ~/.gnupg is actually still required for use-cases where
smartcards are involved [2], but this is not relevant for me.

[0] https://support.mozilla.org/en-US/kb/openpgp-thunderbird-howto-and-faq
[1] https://support.mozilla.org/en-US/kb/openpgp-thunderbird-howto-and-faq#w_where-does-thunderbird-store-openpgp-keys
[2] https://support.mozilla.org/en-US/kb/openpgp-thunderbird-howto-and-faq#w_i-need-to-use-both-gnupg-and-thunderbird-in-parallel-can-i-synchronize-my-keys