Fix(zsh): find command doesn't return sorted list
I realized just now that find return an unsorted list. This cause
aliases to be loaded in random order, which in turn overwrite aliases
that should not be overwritten, like ls and exa.

The order was already decided via filenames (eg. 00-, 10-), so simply
sourcing them in a sorted order fixed the problem.