fix(flatpak): Re-add Thunderbird notification permission
Since libnotify v0.8+ notifications use desktop portal notification [0].
Because of that TB flathub maintainer removed the permission to talk to
org.freedesktop.Notifications [1]. That's fine and works properly if
you only use TB default notifications. In my case I'm using a
bash-port of notify-send [2] in conjuntion with the Mailbox-Alert
plugin [3]. So, I need to add the permission back again.

Refs:
[0] https://gitlab.gnome.org/GNOME/libnotify/-/commit/f3eb807464e06eb68b2901e553a21e31620aba53
[1] https://github.com/flathub/org.mozilla.Thunderbird/commit/64c3abde49b92f00483e90d6ef544aba71b478f4
[2] https://github.com/StayPirate/dotfiles/commit/fd95f630a9132be5437cee7ec40c89e6eb0e9bd6
[3] https://github.com/tjeb/Mailbox-Alert