fix(sway): Hide lsmod errors during startup
As explained here [0], if lsmod is executed while a kernel module is
loading or unloading then an error message is printed by lsmod.
Since I only use lsmod during zsh initialization to detect if the host
system uses a Nvidia GPU or not, there's no risk to hide such errors.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1616232