fix(minicom): Do not use sudo, add the user to uucp instead
I prefer to add my user to the uucp group. This way files created by the
user who run minicon (root with sudo) are created by the same user. I
had a problem with minicon.log which was created by the root user in the
same folder where the command was ran.