feat(git): disable security guard for CVE-2022-24765 [0]
This (now disabled) security feature is meant to check if the user is
also the owner of the local repository folder. That's useful to
prevent users from being compromised in a multi-user system when they
cd into a repository owned by another user and run git commands.

That's not part of my threat model and is just annoying as it
requires manually whitelisting each impacted repository.

  fatal: detected dubious ownership in repository at '...'

Since I sometimes work on sshfs-mounted remote repository, that would
trigger the fatal above.

[0] https://nvd.nist.gov/vuln/detail/CVE-2022-24765